PET-236 refactor : 알림 전송 로직 분리 및 알림 데이터 저장 비동기 방식에서 동기 방식으로 수정
